parsing부터 좌표계 변환까지 완료: 정리 및 개선 필요....
